# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Forge API",
    "description": "Forge backend — AI page builder (see docs/plan/02_PRD.md).",
    "contact": {
      "name": "Forge",
      "url": "http://localhost:3000/"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "development"
    }
  ],
  "paths": {
...
    "/api/v1/studio/generate": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Studio Generate",
        "operationId": "studio_generate_api_v1_studio_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudioGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studio/refine": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Studio Refine",
        "operationId": "studio_refine_api_v1_studio_refine_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudioRefineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studio/sections/edit": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Studio Section Edit",
        "operationId": "studio_section_edit_api_v1_studio_sections_edit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudioSectionEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioSectionEditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studio/conversations/{page_id}": {
      "get": {
        "tags": [
          "studio"
        ],
        "summary": "Get Conversation",
        "operationId": "get_conversation_api_v1_studio_conversations__page_id__get",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Page Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studio/conversations/{page_id}/messages": {
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Post Message",
        "operationId": "post_message_api_v1_studio_conversations__page_id__messages_post",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Page Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudioMessageCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioMessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StudioConversationResponse": {
        "properties": {
          "page_id": {
            "type": "string",
            "format": "uuid",
            "title": "Page Id"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/StudioMessageOut"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "page_id",
          "conversation_id",
          "messages"
        ],
        "title": "StudioConversationResponse"
      },
      "StudioGenerateRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "examples": [
              "Small jobs booking page for Reds Construction"
            ]
          },
          "page_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Id",
            "examples": [
              null
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "gemini"
            ],
            "title": "Provider",
            "default": "openai"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "StudioGenerateRequest"
      },
      "StudioMessageCreateRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "examples": [
              "user"
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "examples": [
              "Add a phone field"
            ]
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "StudioMessageCreateRequest"
      },
      "StudioMessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "title": "StudioMessageOut"
      },
      "StudioRefineRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Make the hero more minimal"
            ]
          },
          "page_id": {
            "type": "string",
            "format": "uuid",
            "title": "Page Id",
            "examples": [
              "00000000-0000-4000-8000-000000000002"
            ]
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "openai",
                  "anthropic",
                  "gemini"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "message",
          "page_id"
        ],
        "title": "StudioRefineRequest"
      },
      "StudioSectionEditRequest": {
        "properties": {
          "page_id": {
            "type": "string",
            "format": "uuid",
            "title": "Page Id",
            "examples": [
              "00000000-0000-4000-8000-000000000002"
            ]
          },
          "section_id": {
            "type": "string",
            "title": "Section Id",
            "examples": [
              "hero-centered-0"
            ]
          },
          "html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html",
            "description": "Optional; server extracts from page when omitted."
          },
          "instruction": {
            "type": "string",
            "title": "Instruction",
            "examples": [
              "Change headline to Hello"
            ]
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "openai",
                  "anthropic",
                  "gemini"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "page_id",
          "section_id",
          "instruction"
        ],
        "title": "StudioSectionEditRequest"
      },
      "StudioSectionEditResponse": {
        "properties": {
          "current_html": {
            "type": "string",
            "title": "Current Html"
          }
        },
        "type": "object",
        "required": [
          "current_html"
        ],
        "title": "StudioSectionEditResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}